+++ CHANGELOG.md
@@ -11,8 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - IScreenFader interface
 - AnimationScreenFader component and prefab
 - ScreenFaderPool
-- ISceneTransitionData, SceneTransitionData and ScriptableSceneTransitionData
 - SceneTransition
+- SceneTransitionData
 
 ### Changed
 - Scriptable Settings Provider for Scene Manager Setting