+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2022-09-03
 ### Added
 - Awaitable Coroutines package 1.0.0
 - Scriptable Settings Provider package 1.0.0
@@ -49,6 +51,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - README
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/SceneManagement/compare/1.0.1...main
+[Unreleased]: https://github.com/HyagoOliveira/SceneManagement/compare/2.0.0...main
+[2.0.0]: https://github.com/HyagoOliveira/SceneManagement/tree/2.0.0/
 [1.0.1]: https://github.com/HyagoOliveira/SceneManagement/tree/1.0.1/
 [1.0.0]: https://github.com/HyagoOliveira/SceneManagement/tree/1.0.0/